refactored to reduce the size of main()
I think that main() in executable.go should focus just on the behavior
of the program. There's no need for a casual reader to see the derping
that is required to display a simple cross-shaped window under X11.

As such I have relocated many functions from executable.go to util/*,
and factored the creation of the cross UI into ui/cross.go, which is a
somewhat more sensible place for it.