fixed mysterious fluxbox vertical resize-on-move bug
The solution is a brute-force one: if a window changes size after a move
operation, we just change it back. To do so, j3 now uses new functions
`wm.Move` and `wm.MoveResize` instead of the corresponding methods on
`xwindow.Window`. This is not a pretty solution, but it (seems to) work
for the most part, and I don't want to fix bugs in Fluxbox!

As part of these changes, resize.go lost the `WaitForGeometryChange`
function, which was moved into wm/moveresize.go, and changed to a
synchronous form.